Move the customized hook "useUser" to hooks/
